+++ tubular/mixins.py
@@ -4,7 +4,6 @@ from typing import TYPE_CHECKING
 
 import narwhals as nw
 import narwhals.selectors as ncs
-import numpy as np
 
 if TYPE_CHECKING:
     import pandas as pd
@@ -174,8 +173,8 @@ class WeightColumnMixin:
             msg = f"{self.classname()}: weight column must be non-null"
             raise ValueError(msg)
 
-        # check weight not inf, currently no polars-y way to do this in narwhals
-        if np.isinf(X[weights_column].to_numpy()).any():
+        # check weight not inf
+        if not X.select((nw.col(weights_column).is_finite()).all()).item():
             msg = f"{self.classname()}: weight column must not contain infinite values."
             raise ValueError(msg)